==> notebooks/02_click_number_predictor_model_training.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import nltk

%matplotlib inline

#Example file to see what kind of data we are working with
df_example = pd.read_excel('articles_example_2024.xlsx')
df_example

df = pd.read_excel('articles_2024.xlsx')

df = df[2:-17] #Cutting off articles from 2023 and 2025

df.drop('URL', axis = 1, inplace = True) #Drop URL column

df = df[df['Author'] != 'Automatic'] #Drop automatic articles
df.reset_index(inplace = True, drop = True)

#For some reason there is a 'No heading' category, we will rename it
df[df['Heading'] == 'No heading']
df.iloc[6373, 4] = 'Savings'

df = pd.get_dummies(df, columns = ['Lead article?'], drop_first = False) #Grouping leaders into separate columns with dummy variables
df.iloc[:,-5:] = df.iloc[:,-5:].astype(int) #Convert True and False values in new columns to numbers

df.drop(['Date', 'Robot?', 'Heading', 'Author', 'Top'], axis = 1, inplace = True) #Drop unnecessary columns

df['Cleaned_title'] = df['Title'].str.lower() #Lowercase titles

#Special characters need to be removed
import re

df['Cleaned_title'] = df['Cleaned_title'].apply(lambda x: re.sub(r'[^a-záéíóöőúüű0-9\s]', '', x)) #Removing missing characters from []
df['Cleaned_title'] = df['Cleaned_title'].str.replace(r'\s+', ' ', regex = True).str.strip() #Merge multiple consecutive spaces into one space

#Removing stop words (e.g. 'and', 'a', 'an', 'the')

from nltk.corpus import stopwords

#Adding unique stop words
unique_stop_words = ['is', 'ha', 'le', 'fog', '2024', '2024ben', 'te', 'előtt', 'dolog', 'nap', 'párizs', 'kvíz', 'mire', 'olimpia']

stop_words = set(stopwords.words('hungarian'))
stop_words.update(unique_stop_words)
df['Cleaned_title'] = df['Cleaned_title'].apply(lambda x: ' '.join([word for word in x.split() if word not in stop_words]))

df.reset_index(inplace = True, drop = True)

from collections import Counter

all_words = ' '.join(df['Cleaned_title']).split() #Concatenate all titles into a single string with spaces, then split them
                                                   #along the spaces, putting the words into one big list
word_counts = Counter(all_words)
print(word_counts.most_common(10))

all_words = list(set(all_words))
#all_words

len(all_words)

#LET'S LOOK AT THE MODEL ITSELF

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error

#Extract the line(s) containing the longest text
max_length = df['Cleaned_title'].apply(len).max()

vectorizer = TfidfVectorizer(max_features = 200)
X_text = vectorizer.fit_transform(df['Cleaned_title']).toarray()

X_features = df.iloc[:, 2:-1]
X = pd.concat([pd.DataFrame(X_text), X_features], axis = 1) #Convert X_text to DF and merge it with other features

y = df['Stat']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

#Convert column names to strings (this is needed for RandomForestRegressor)
X_train.columns = X_train.columns.astype(str)
X_test.columns = X_test.columns.astype(str)

print(X_train.dtypes)

model = RandomForestRegressor()
model.fit(X_train, y_train)

y_pred = model.predict(X_test)
rmse = mean_squared_error(y_test, y_pred, squared = False)
print(f'RMSE: {rmse}')

y_pred

#TESTING

probe = pd.DataFrame({
    'Title': ['Őrült árvágtában a magyarok kedvenc gyümölcslevei: mindenki nézhet egy nagyot a kasszáknál'], #Example article title
    'Lead article?_Afternoon lead article': [1],
    'Lead article?_Early morning lead article': [0],
    'Lead article?_Morning lead article': [0],
    'Lead article?_Not lead article': [0],
    'Lead article?_Tabloid lead article': [0]
}
                    )

probe

probe['Cleaned_title'] = probe['Title'].str.lower() #Lowercase titles

probe['Cleaned_title'] = probe['Cleaned_title'].apply(lambda x: re.sub(r'[^a-záéíóöőúüű0-9\s]', '', x)) #Removing missing characters from []
probe['Cleaned_title'] = probe['Cleaned_title'].str.replace(r'\s+', ' ', regex = True).str.strip() #Merge multiple consecutive spaces into one space

probe['Cleaned_title'] = probe['Cleaned_title'].apply(lambda x: ' '.join([word for word in x.split() if word not in stop_words]))

#vectorizer = TfidfVectorizer(max_features = 500)
probe_text = vectorizer.transform(probe['Cleaned_title']).toarray()

probe = pd.concat([pd.DataFrame(probe_text), probe], axis = 1) ##Convert probe_text to DF and merge it with other features

probe.drop(['Title', 'Cleaned_title'], inplace = True, axis = 1)

#Convert column names to strings (this is needed for RandomForestRegressor)
probe.columns = probe.columns.astype(str)

predicted_clicks = model.predict(probe)

print(predicted_clicks)

#Save the model
import joblib
joblib.dump(model, 'readership_predictor.pkl')

#Use saved model
loaded_model = joblib.load('readership_predictor.pkl')
prediction = loaded_model.predict(probe)
print(prediction)

#Save the trained vectorizer
joblib.dump(vectorizer, 'tfidf_vectorizer.pkl')
