==> notebooks/03_load_and_use_the_model.ipynb <==
import joblib
import pandas as pd
import re
import sys
sys.path.append('../src') #Add src/ to the import path
from utils import preprocess_article

loaded_model = joblib.load('readership_predictor.pkl')

probe_article_title = preprocess_article('Devizapiac: vegyesen mozgott a forint reggel',
                                   'Not lead') #Probe article title

predicted_clicks = loaded_model.predict(probe_article_title)
print(predicted_clicks)
